% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/robusta.R
\name{robust}
\alias{robust}
\title{Specifying the type of robust standard error to be reported in regression summary}
\usage{
robust(reg, se = c("Hinkley", "hc1", "stata", "White", "hc0", "HornDuncan",
  "hc2", "Efron", "hc3", "CribariNeto", "hc4", "NeweyWest", "Andrews", "hac",
  "Lumley", "Arellano"))
}
\arguments{
\item{reg}{regression objects, for now only works for lm, glm, and plm objects}

\item{se}{type of robust standard error. The default is HC1 to make it easy to compare it with stata robust standard error. The default heteroskedasticity consistent (HC) covariance matrix in sandwich package is instead HC3, that is less affected by outlier in small sample. All types of robust standard error can be referred by the name of author who propose it. 'hc1' is equivalent to 'Hinkley', 'hc0' is equivalent to 'White', 'hc2' is equivalent to 'Horn', 'hc3' is equivalent to 'Efron', 'hc4' is equivalent to 'Cribari'. For data containing serial correlation, the available heteroscedasticity and autocorrelation consistent (HAC) standard error is 'NeweyWest', 'Andrew' that is equivalent to 'hac', and 'Lumley'. All HC types and 'arrellano' method are available for plm object in panel data analysis.}
}
\description{
Summary command will identify which type of standard error to use by reading robust attribute attached by this command
}
\examples{
lmo = lm(mpg ~ carb + cyl, mtcars)
texreg::screenreg(list(lmo, robust(lmo), robust(lmo, 'hc3'), robust(lmo, 'White')))
}

